Renamed libcrow to stdlibcrow to be more explicit.
