+++ cmake/targets.cmake
@@ -1,17 +1,17 @@
 # Set the different CMake targets.
 
 # Define the possible targets/builds.
-set(TARGET_CROW_BIN "crow")
-set(TARGET_CROW_LIB "crowlib")
+set(TARGET_CROW_BIN "crow")    # Compiler binary.
+set(TARGET_CROW_LIB "crowlib") # Compiler library.
 
-set(TARGET_CROW_STDLIB "stdcrowlib")
+set(TARGET_CROW_STDLIBCROW "stdlibcrow") # Standard library for generated C++.
 
-set(TARGET_CROW_TESTS "crow_tests")
+set(TARGET_CROW_TESTS "crow_tests") # Unit test binary
 
 # Conditionally set the main target:
+# Change the main target if we are building tests:
 set(TARGET_CROW "crow")
 
-# Change the main target if we are building tests.
 if(DEFINED CROW_BUILD_TESTS)
 set(TARGET_CROW ${TARGET_CROW_TESTS})
 endif()
