+++ src/CMakeLists.txt
@@ -5,4 +5,4 @@ add_subdirectory(crow)
 add_subdirectory(lib)
 
 # Standard library for Crow.
-add_subdirectory(libcrow)
+add_subdirectory(stdlibcrow)
